traemos producto de bd, deberiamos agregar mas a bd, pero la pagina de products ya estaria lista, quizas algunos estilos a pulir. estamos avanzando en click en producto para ir a singleProduct
