+++ components/ProductCard/ProductCard.js
@@ -5,9 +5,9 @@ import styles from './productCard.module.css';
 const ProductCard = ({ imagen, nombre, precio }) => {
   return (
     <div className={styles['productCard']}>
-      <Image src={imagen} alt={nombre}></Image>
+      <Image src={imagen} alt={nombre} width={'228px'} height={'228px'}></Image>
       <p>{nombre}</p>
-      <p className={styles['precio']}>{precio}</p>
+      <p className={styles['precio']}>{`$ ${precio}`}</p>
     </div>
   );
 };
