+++ pages/index.js
@@ -10,9 +10,7 @@ import { useEffect, useState, useContext } from 'react';
 import CardFeature from '../components/CardFeature/CardFeature';
 import { faLightbulb, faSmile } from '@fortawesome/free-regular-svg-icons';
 import { faWind, faLeaf } from '@fortawesome/free-solid-svg-icons';
-import { CartContext } from '../contexts/CartContext';
-import { getFirestore } from '../utils/firebase';
-import { doc } from 'prettier';
+//import { doc } from 'prettier';
 
 const dataItems = [
   {
@@ -42,40 +40,11 @@ const dataItems = [
 ];
 
 export default function Home() {
-  const { cart, setCart } = useContext(CartContext);
   const [cardsFeature, setCardsFeature] = useState(['Destacados', 'Nuevos Bonsais', 'Descuentos']);
 
   const showCardsFeature = () => {
     return cardsFeature.map((card, index) => <CardFeature key={index} text={card} />);
   };
-
-  // useEffect(() => {
-  //   setCart([
-  //     { id: 1, name: 'Robert' },
-  //     { id: 2, name: 'Mike' },
-  //   ]);
-  // }, [setCart]);
-
-  useEffect(() => {
-    const getProducts = async () => {
-      try {
-        const db = getFirestore();
-        const itemsCollection = db.collection(`accesorios`);
-        console.log(itemsCollection);
-        const itemSnapshot = await itemsCollection.get();
-
-        const items = itemSnapshot.docs.map((doc) => {
-          return { id: doc.id, ...doc.data() };
-        });
-
-        console.log(items);
-      } catch (error) {
-        console.log(error);
-      }
-    };
-
-    getProducts();
-  }, []);
 
   const dataPage = { page: 'index' };
